install tslint
