+++ .vscode/extensions.json
@@ -1,0 +1,5 @@
+{
+  "recommendations": [
+    "ms-vscode.vscode-typescript-tslint-plugin"
+  ]
+}
