+++ .vscode/settings.json
@@ -1,5 +1,6 @@
 {
   "editor.codeActionsOnSave": {
+    "source.fixAll.tslint": true,
     "source.fixAll.eslint": true,
     "source.fixAll.stylelint": true
   },
