+++ package.json
@@ -43,6 +43,7 @@
     "stylelint-scss": "^3.19.0",
     "stylelint-selector-bem-pattern": "^2.1.0",
     "ts-loader": "^8.1.0",
+    "tslint": "^6.1.3",
     "typescript": "^4.2.4",
     "webpack": "^5.31.2",
     "webpack-cli": "^4.6.0",
