+++ tslint.json
@@ -1,0 +1,9 @@
+{
+  "rules": {
+    "indent": [true, "spaces", 2],
+    "linebreak-style": [true, "LF"],
+    "object-literal-key-quotes": [true, "as-needed"],
+    "type-literal-delimiter": true,
+    "triple-equals": [true, "allow-null-check"]
+  }
+}
